+++ packages/gamut/Markdown/index.tsx
@@ -8,16 +8,17 @@ import {
   createTagOverride,
   createCodeBlockOverride,
   ManyOverrideSettings,
+  standardOverrides,
 } from './libs/overrides';
 import s from './styles/index.scss';
-
 import Iframe from './overrides/Iframe';
 import Anchor from './overrides/Anchor';
 
-const htmlToReactParser = new HtmlToReact.Parser();
-const processNodeDefinitions = new HtmlToReact.ProcessNodeDefinitions();
+const htmlToReactParser = new HtmlToReact.Parser({
+  xmlMode: true,
+});
 
-const sanitizationConfig = {
+const defaultSanitizationConfig = {
   allowedAttributes: {
     ...insane.defaults.allowedAttributes,
     span: ['class'],
@@ -97,24 +98,38 @@ class Markdown extends PureComponent<MarkdownProps> {
         component: Anchor,
       }),
       ...overrides,
-      {
-        shouldProcessNode() {
-          return true;
-        },
-        processNode: processNodeDefinitions.processDefaultNode,
-      },
+      ...standardOverrides,
     ];
 
-    // Render markdown to html
-    const rawHtml = inline ? marked.inlineLexer(text, []) : marked(text);
+    const markedOptions = {
+      smartypants: true,
+    };
 
-    const html = insane(rawHtml, {
-      ...sanitizationConfig,
+    // Render markdown to html
+    const rawHtml = inline
+      ? marked.inlineLexer(text, [], markedOptions)
+      : marked(text, markedOptions);
+
+    const sanitizationConfig = {
+      ...defaultSanitizationConfig,
       allowedTags: [
-        ...sanitizationConfig.allowedTags,
-        ...Object.keys(userOverrides).map(key => key.toLowerCase()),
+        ...defaultSanitizationConfig.allowedTags,
+        ...Object.keys(userOverrides).map(tagName => tagName.toLowerCase()),
       ],
-    });
+      allowedAttributes: {
+        ...defaultSanitizationConfig.allowedAttributes,
+        ...Object.keys(userOverrides).reduce((acc, tagName) => {
+          return {
+            ...acc,
+            [tagName.toLowerCase()]: (
+              userOverrides[tagName].allowedAttributes || []
+            ).map(attr => attr.toLowerCase()),
+          };
+        }, {}),
+      },
+    };
+
+    const html = insane(rawHtml, sanitizationConfig);
 
     // Render html to a react tree
     const react = htmlToReactParser.parseWithInstructions(