+++ packages/gamut/Markdown/libs/overrides.tsx
@@ -1,9 +1,12 @@
 import React from 'react';
 import { get } from 'lodash';
+import HtmlToReact from 'html-to-react';
 import camelCaseMap from 'html-to-react/lib/camel-case-attribute-names';
 
+const processNodeDefinitions = new HtmlToReact.ProcessNodeDefinitions();
+
 export interface AttributesMap {
-  [key: string]: string;
+  [key: string]: string | boolean;
 }
 
 export interface HTMLToReactNode {
@@ -18,7 +21,7 @@ export interface HTMLToReactNode {
 }
 
 // Mapping of html attributes to their camelCase variants
-const attributeMap: AttributesMap = {
+const attributeMap: { [key: string]: string } = {
   ...camelCaseMap,
   for: 'htmlFor',
   class: 'className',
@@ -26,6 +29,7 @@ const attributeMap: AttributesMap = {
 
 export type OverrideSettings = {
   component: React.ComponentType;
+  allowedAttributes?: string[];
   processNode?: (node: HTMLToReactNode, props: object) => React.ReactNode;
   shouldProcessNode?: (node: HTMLToReactNode) => boolean;
 };
@@ -34,13 +38,26 @@ export type ManyOverrideSettings = {
   [i: string]: OverrideSettings;
 };
 
+const processAttributeValue = (value: string | boolean) => {
+  if (value === 'true') {
+    return true;
+  }
+
+  if (value === 'false') {
+    return false;
+  }
+
+  return value || true;
+};
+
 // Convert html attributes to valid react props
 export const processAttributes = (attributes: AttributesMap = {}) =>
   Object.keys(attributes).reduce((acc, attr) => {
-    const key = attributeMap[attr.replace(/[-:]/, '')];
+    const reactAttr = attributeMap[attr.replace(/[-:]/, '')] || attr;
+
     return {
       ...acc,
-      [key || attr]: attributes[attr],
+      [reactAttr]: processAttributeValue(attributes[attr]),
     };
   }, {});
 
@@ -69,9 +86,11 @@ export const createTagOverride = (
       children,
       key,
     };
+
     if (Override.processNode) {
       return Override.processNode(node, props);
     }
+
     return <Override.component {...props} />;
   },
 });
@@ -101,3 +120,30 @@ export const createCodeBlockOverride = (
     },
     ...Override,
   });
+
+const processText = (text: string) => {
+  // Replace &mdash; due to legacy markdown that didn't use smart dashes
+  return text.replace(/&mdash;/g, '\u2014');
+};
+
+export const standardOverrides = [
+  {
+    shouldProcessNode(node: HTMLToReactNode) {
+      // Parse text outside of code blocks
+      if (node.parent && ['code', 'pre'].indexOf(node.parent.name) >= 0) {
+        return false;
+      }
+      if (node.type === 'text') return true;
+      return false;
+    },
+    processNode(node: HTMLToReactNode, ...args: any[]) {
+      return processText(node.data);
+    },
+  },
+  {
+    shouldProcessNode() {
+      return true;
+    },
+    processNode: processNodeDefinitions.processDefaultNode,
+  },
+];