+++ packages/gamut-styles/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.2.0](https://github.com/RyzacInc/client-modules/compare/@codecademy/gamut-styles@2.1.0...@codecademy/gamut-styles@2.2.0) (2019-03-13)
+
+
+### Features
+
+* Add shared config modules for eslint & prettier ([#224](https://github.com/RyzacInc/client-modules/issues/224)) ([7fbeac6](https://github.com/RyzacInc/client-modules/commit/7fbeac6))
+
+
+
+
+
+# [2.1.0](https://github.com/RyzacInc/gamut/compare/@codecademy/gamut-styles@2.0.4...@codecademy/gamut-styles@2.1.0) (2019-03-13)
+
+
+### Features
+
+* Adds mar 2019 colors, deprecates colors ([#211](https://github.com/RyzacInc/gamut/issues/211)) ([c5c1e7f](https://github.com/RyzacInc/gamut/commit/c5c1e7f))
+
+
+
+
+
 <a name="2.0.4"></a>
 ## [2.0.4](https://github.com/RyzacInc/gamut/compare/@codecademy/gamut-styles@2.0.3...@codecademy/gamut-styles@2.0.4) (2019-01-16)